+++ CHANGELOG.md
@@ -14,9 +14,13 @@
 
 ### Changes
 - Stop sending `X-Neptune-LegacyClient` header ([#1715](https://github.com/neptune-ai/neptune-client/pull/1715))
+- Use `tqdm.auto` ([#1717](https://github.com/neptune-ai/neptune-client/pull/1717))
 
 ### Features
 - ?
+
+### Fixes
+- Fixed `tqdm.notebook` import only in Notebook environment ([#1716](https://github.com/neptune-ai/neptune-client/pull/1716))
 
 
 ## neptune 1.10.0